+++ tests/test_models.py
@@ -42,36 +42,14 @@ class TestModel:
         # Check representation of Model.
         assert 'Model [resistivity]; isotropic' in model1.__repr__()
 
-        model1e = models.Model(grid)
-        model1.mu_r = None
-        model1.epsilon_r = None
         vmodel1 = models.VolumeModel(model1, sfield)
-        assert_allclose(model1.property_x, model1.property_y)
         assert_allclose(model1.size, grid.n_cells)
         assert_allclose(model1.shape, grid.shape_cells)
         assert_allclose(vmodel1.eta_z, vmodel1.eta_y)
+        assert model1.property_y is None
+        assert model1.property_z is None
         assert model1.mu_r is None
         assert model1.epsilon_r is None
-
-        # Assert that the cases changes if we assign y- and z- resistivities
-        # (I).
-        model1.property_x = model1.property_x
-        assert model1.case == 0
-        model1.property_y = 2*model1.property_x
-        assert model1.case == 1
-        model1.property_z = 2*model1.property_x
-        assert model1.case == 3
-
-        assert model1e.case == 0
-        model1e.property_z = 2*model1.property_x
-        assert model1e.case == 2
-        model1e.property_y = 2*model1.property_x
-        assert model1e.case == 3
-
-        model1.epsilon_r = 1  # Just set to 1 CURRENTLY.
-        vmodel1b = models.VolumeModel(model1, sfield)
-        assert_allclose(vmodel1b.eta_x, vmodel1.eta_x)
-        assert np.iscomplex(vmodel1b.eta_x[0, 0, 0])
 
         # Using ints
         model2 = models.Model(grid, 2., 3., 4.)
@@ -80,9 +58,9 @@ class TestModel:
         assert_allclose(model2.property_x*2, model2.property_z)
 
         # VTI: Setting property_x and property_z, not property_y
-        model2b = models.Model(grid, 2., property_z=4.)
+        model2b = models.Model(grid, 2., property_z=4., epsilon_r=1)
         vmodel2b = models.VolumeModel(model2b, sfield)
-        assert_allclose(model2b.property_x, model2b.property_y)
+        assert model1.property_y is None
         assert_allclose(vmodel2b.eta_y, vmodel2b.eta_x)
         model2b.property_z = model2b.property_x
         model2c = models.Model(grid, 2., property_z=model2b.property_z.copy())
@@ -93,7 +71,7 @@ class TestModel:
         # HTI: Setting property_x and property_y, not property_z
         model2d = models.Model(grid, 2., 4.)
         vmodel2d = models.VolumeModel(model2d, sfield)
-        assert_allclose(model2d.property_x, model2d.property_z)
+        assert model1.property_z is None
         assert_allclose(vmodel2d.eta_z, vmodel2d.eta_x)
 
         # Pure air, epsilon_r init with 0 => should be 1!
@@ -101,14 +79,13 @@ class TestModel:
         assert model6.epsilon_r is None
 
         # Check wrong shape
-        with pytest.raises(ValueError, match='Shape of property_x must be ()'):
+        with pytest.raises(ValueError, match='could not be broadcast'):
             models.Model(grid, np.arange(1, 11))
-        with pytest.raises(ValueError, match='Shape of property_y must be ()'):
+        with pytest.raises(ValueError, match='could not be broadcast'):
             models.Model(grid, property_y=np.ones((2, 5, 6)))
-        with pytest.raises(ValueError, match='Shape of property_z must be ()'):
-            models.Model(grid, property_z=np.array([1, 3]))
-        with pytest.raises(ValueError, match='Shape of mu_r must be ()'):
-            models.Model(grid, mu_r=np.array([[1, ], [3, ]]))
+        # Actual broadcasting.
+        models.Model(grid, property_z=np.array([1, 3]))
+        models.Model(grid, mu_r=np.array([[1, ], [3, ]]))
 
         # Check with all inputs
         gridvol = grid.cell_volumes.reshape(grid.shape_cells, order='F')
@@ -136,11 +113,9 @@ class TestModel:
         model3.property_y[:, 1:, :] = tres[:, 1:, :]*3.0
         model3.property_z = tres*4.0
         model3.mu_r = tres*5.0
-        model3.epsilon_r = tres*6.0
         assert_allclose(tres*2., model3.property_x)
         assert_allclose(tres*3., model3.property_y)
         assert_allclose(tres*4., model3.property_z)
-        assert_allclose(tres*6., model3.epsilon_r)
 
         # Check eta
         iomep = sfield.sval*models.epsilon_0
@@ -165,9 +140,9 @@ class TestModel:
             _ = models.Model(grid, property_x=property_x*0)
         Model = models.Model(grid, property_x=property_x)
         with pytest.raises(ValueError, match='`property_x` must be all'):
-            Model._check_parameter(property_x*0, 'property_x')
+            Model._check_positive_finite(property_x*0, 'property_x')
         with pytest.raises(ValueError, match='`property_x` must be all'):
-            Model._check_parameter(-1.0, 'property_x')
+            Model._check_positive_finite(-1.0, 'property_x')
         with pytest.raises(ValueError, match='`property_y` must be all'):
             _ = models.Model(grid, property_y=np.inf)
         with pytest.raises(ValueError, match='`property_z` must be all'):
@@ -194,32 +169,29 @@ class TestModel:
 
         model1inp = models.Model(grid, property_x)
 
-        model1out = model1inp.interpolate2grid(grid, grid2)
-        assert_allclose(model1out._property_x[0],
+        model1out = model1inp.interpolate2grid(grid2)
+        assert_allclose(model1out.property_x[0],
                         10**(np.sum(np.log10(model1inp.property_x))/8))
-        assert model1out._property_y is None
-        assert model1out._property_z is None
-        assert model1out._mu_r is None
-        assert model1out._epsilon_r is None
+        assert model1out.property_y is None
+        assert model1out.property_z is None
+        assert model1out.epsilon_r is None
+        assert model1out.mu_r is None
 
         model2inp = models.Model(
                 grid, property_x=property_x, property_y=property_y,
                 property_z=property_z, mu_r=mu_r, epsilon_r=epsilon_r)
 
-        model2out = model2inp.interpolate2grid(grid, grid2)
-        assert_allclose(model2out._property_x[0],
+        model2out = model2inp.interpolate2grid(grid2)
+        assert_allclose(model2out.property_x[0],
                         10**(np.sum(np.log10(model2inp.property_x))/8))
-        assert_allclose(model2out._property_y[0],
+        assert_allclose(model2out.property_y[0],
                         10**(np.sum(np.log10(model2inp.property_y))/8))
-        assert_allclose(model2out._property_z[0],
+        assert_allclose(model2out.property_z[0],
                         10**(np.sum(np.log10(model2inp.property_z))/8))
-        assert_allclose(model2out._mu_r[0],
-                        10**(np.sum(np.log10(model2inp.mu_r))/8))
-        assert_allclose(model2out._epsilon_r[0],
+        assert_allclose(model2out.epsilon_r,
                         10**(np.sum(np.log10(model2inp.epsilon_r))/8))
-
-
-class TestModel2:
+        assert_allclose(model2out.mu_r,
+                        10**(np.sum(np.log10(model2inp.mu_r))/8))
 
     def test_kwargs(self):
 
@@ -263,6 +235,49 @@ class TestModel2:
         models.Model(grid, property_x=-1, mapping='LgResistivity')
         models.Model(grid, property_x=-1, mapping='LnResistivity')
 
+    def test_broadcasting(self):
+        grid = meshes.TensorMesh(
+                [np.ones(4), np.ones(2), np.ones(6)], (0, 0, 0))
+
+        model = models.Model(grid, 1, 1, 1, 1, 1)
+        assert_allclose(model.property_x, model.property_y)
+        assert_allclose(model.property_x, model.property_z)
+        assert_allclose(model.property_x, model.mu_r)
+        assert_allclose(model.property_x, model.epsilon_r)
+
+        model.property_x = [[[1.]], [[2.]], [[3.]], [[4.]]]
+        assert_allclose(model.property_x[:, 0, 0], [1, 2, 3, 4])
+
+        model.property_y = [[[1, ], [2, ]]]
+        assert_allclose(model.property_y[0, :, 0], [1, 2])
+
+        model.property_z = [[[1, 2., 3., 4., 5., 6.]]]
+        assert_allclose(model.property_z[0, 0, :], [1, 2, 3, 4, 5, 6])
+
+        model.mu_r = 3.33
+        assert_allclose(model.mu_r, 3.33)
+
+        data = np.arange(1, 4*2*6+1).reshape((4, 2, 6), order='F')
+        model.epsilon_r = data
+        assert_allclose(model.epsilon_r, data)
+
+    def test_change_anisotropy(self):
+        hx = [2, 2]
+        grid = meshes.TensorMesh([hx, hx, hx], (0, 0, 0))
+        model = models.Model(grid, 1)
+
+        with pytest.raises(ValueError, match='initiated without `property_y`'):
+            model.property_y = 3
+
+        with pytest.raises(ValueError, match='initiated without `property_z`'):
+            model.property_z = 3
+
+        with pytest.raises(ValueError, match='l was initiated without `mu_r`'):
+            model.mu_r = 3
+
+        with pytest.raises(ValueError, match=' initiated without `epsilon_r`'):
+            model.epsilon_r = 3
+
 
 class TestModelOperators:
 
@@ -273,40 +288,43 @@ class TestModelOperators:
             [np.ones(3), np.ones(4), np.ones(6)], origin=np.array([0, 0, 0]))
 
     # Define a couple of models.
-    model_int = models.Model(mesh_base, 1.)
-    model_1_a = models.Model(mesh_base, 1., 2.)
-    model_1_b = models.Model(mesh_base, 2., 4.)
-    model_2_a = models.Model(mesh_base, 1., property_z=3.)
-    model_2_b = models.Model(mesh_base, 2., property_z=6.)
-    model_3_a = models.Model(mesh_base, 1., 2., 3.)
-    model_3_b = models.Model(mesh_base, 2., 4., 6.)
-    model_mu_a = models.Model(mesh_base, 1., mu_r=1.)
-    model_mu_b = models.Model(mesh_base, 2., mu_r=2.)
-    model_epsilon_a = models.Model(mesh_base, 1., epsilon_r=1.)
-    model_epsilon_b = models.Model(mesh_base, 2., epsilon_r=2.)
+    mod_iso = models.Model(mesh_base, 1.)
+    mod_map = models.Model(mesh_base, 1., mapping='Conductivity')
+    mod_hti_a = models.Model(mesh_base, 1., 2.)
+    mod_hti_b = models.Model(mesh_base, 2., 4.)
+    mod_vti_a = models.Model(mesh_base, 1., property_z=3.)
+    mod_vti_b = models.Model(mesh_base, 2., property_z=6.)
+    mod_tri_a = models.Model(mesh_base, 1., 2., 3.)
+    mod_tri_b = models.Model(mesh_base, 2., 4., 6.)
+    mod_mu_a = models.Model(mesh_base, 1., mu_r=1.)
+    mod_mu_b = models.Model(mesh_base, 2., mu_r=2.)
+    mod_epsilon_a = models.Model(mesh_base, 1., epsilon_r=1.)
+    mod_epsilon_b = models.Model(mesh_base, 2., epsilon_r=2.)
 
-    model_int_diff = models.Model(mesh_diff, 1.)
-    model_nC = models.Model(mesh_base, np.ones(mesh_base.n_cells)*2.)
-    model_shape_cells = models.Model(mesh_base, np.ones(mesh_base.shape_cells))
+    mod_int_diff = models.Model(mesh_diff, 1.)
+    mod_size = models.Model(mesh_base, np.ones(mesh_base.n_cells)*2.)
+    mod_shape = models.Model(mesh_base, np.ones(mesh_base.shape_cells))
 
     def test_operator_test(self):
-        with pytest.raises(ValueError, match='Models must be of the'):
-            self.model_int + self.model_1_a
-        with pytest.raises(ValueError, match='Models must be of the'):
-            self.model_int - self.model_2_a
-        with pytest.raises(ValueError, match='Models must be of the'):
-            self.model_int + self.model_3_a
-        with pytest.raises(ValueError, match='Either both or none of'):
-            self.model_int - self.model_mu_a
-        with pytest.raises(ValueError, match='Either both or none of'):
-            self.model_int + self.model_epsilon_a
-        with pytest.raises(ValueError, match='Models could not be broadcast'):
-            self.model_int - self.model_int_diff
+        with pytest.raises(ValueError, match='Models have different anisotro'):
+            self.mod_iso + self.mod_hti_a
+        with pytest.raises(ValueError, match='Models have different anisotro'):
+            self.mod_iso - self.mod_vti_a
+        with pytest.raises(ValueError, match='Models have different anisotro'):
+            self.mod_iso + self.mod_tri_a
+        with pytest.raises(ValueError, match='One model has mu_r, the other '):
+            self.mod_iso - self.mod_mu_a
+        with pytest.raises(ValueError, match='One model has epsilon_r, the o'):
+            self.mod_iso + self.mod_epsilon_a
+        with pytest.raises(ValueError, match='Models have different grids.'):
+            self.mod_iso - self.mod_int_diff
+        with pytest.raises(ValueError, match='Models have different mappings'):
+            self.mod_iso - self.mod_map
 
     def test_add(self):
-        a = self.model_int + self.model_nC
-        b = self.model_nC + self.model_shape_cells
-        c = self.model_int + self.model_shape_cells
+        a = self.mod_iso + self.mod_size
+        b = self.mod_size + self.mod_shape
+        c = self.mod_iso + self.mod_shape
         assert a == b
         assert a != c
         assert a.property_x[0, 0, 0] == 3.0
@@ -315,101 +333,101 @@ class TestModelOperators:
 
         # Addition with something else than a model
         with pytest.raises(TypeError):
-            self.model_int + self.mesh_base
+            self.mod_iso + self.mesh_base
 
         # All different cases
-        a = self.model_1_a + self.model_1_a
-        assert a.property_x == self.model_1_b.property_x
-        assert a.property_y == self.model_1_b.property_y
-        assert a.property_z.base is a.property_x.base
+        a = self.mod_hti_a + self.mod_hti_a
+        assert_allclose(a.property_x, self.mod_hti_b.property_x)
+        assert_allclose(a.property_y, self.mod_hti_b.property_y)
+        # assert a.property_z.base is a.property_x.base
 
-        a = self.model_2_a + self.model_2_a
-        assert a.property_x == self.model_2_b.property_x
-        assert a.property_y.base is a.property_x.base
-        assert a.property_z == self.model_2_b.property_z
+        a = self.mod_vti_a + self.mod_vti_a
+        assert_allclose(a.property_x, self.mod_vti_b.property_x)
+        # assert a.property_y.base is a.property_x.base
+        assert_allclose(a.property_z, self.mod_vti_b.property_z)
 
-        a = self.model_3_a + self.model_3_a
-        assert a.property_x == self.model_3_b.property_x
-        assert a.property_y == self.model_3_b.property_y
-        assert a.property_z == self.model_3_b.property_z
+        a = self.mod_tri_a + self.mod_tri_a
+        assert_allclose(a.property_x, self.mod_tri_b.property_x)
+        assert_allclose(a.property_y, self.mod_tri_b.property_y)
+        assert_allclose(a.property_z, self.mod_tri_b.property_z)
 
         # mu_r and epsilon_r
-        a = self.model_mu_a + self.model_mu_a
-        b = self.model_epsilon_a + self.model_epsilon_a
-        assert a.mu_r == self.model_mu_b.mu_r
-        assert b.epsilon_r == self.model_epsilon_b.epsilon_r
+        a = self.mod_mu_a + self.mod_mu_a
+        b = self.mod_epsilon_a + self.mod_epsilon_a
+        assert_allclose(a.mu_r, self.mod_mu_b.mu_r)
+        assert_allclose(b.epsilon_r, self.mod_epsilon_b.epsilon_r)
 
     def test_sub(self):
         # Addition with something else than a model
         with pytest.raises(TypeError):
-            self.model_int - self.mesh_base
+            self.mod_iso - self.mesh_base
 
         # All different cases
-        a = self.model_1_b - self.model_1_a
-        assert a.property_x == self.model_1_a.property_x
-        assert a.property_y == self.model_1_a.property_y
-        assert a.property_z.base is a.property_x.base
+        a = self.mod_hti_b - self.mod_hti_a
+        assert_allclose(a.property_x, self.mod_hti_a.property_x)
+        assert_allclose(a.property_y, self.mod_hti_a.property_y)
+        # assert a.property_z.base is a.property_x.base
 
-        a = self.model_2_b - self.model_2_a
-        assert a.property_x == self.model_2_a.property_x
-        assert a.property_y.base is a.property_x.base
-        assert a.property_z == self.model_2_a.property_z
+        a = self.mod_vti_b - self.mod_vti_a
+        assert_allclose(a.property_x, self.mod_vti_a.property_x)
+        # assert a.property_y.base is a.property_x.base
+        assert_allclose(a.property_z, self.mod_vti_a.property_z)
 
-        a = self.model_3_b - self.model_3_a
-        assert a.property_x == self.model_3_a.property_x
-        assert a.property_y == self.model_3_a.property_y
-        assert a.property_z == self.model_3_a.property_z
+        a = self.mod_tri_b - self.mod_tri_a
+        assert_allclose(a.property_x, self.mod_tri_a.property_x)
+        assert_allclose(a.property_y, self.mod_tri_a.property_y)
+        assert_allclose(a.property_z, self.mod_tri_a.property_z)
 
         # mu_r and epsilon_r
-        a = self.model_mu_b - self.model_mu_a
-        b = self.model_epsilon_b - self.model_epsilon_a
-        assert a.mu_r == self.model_mu_a.mu_r
-        assert b.epsilon_r == self.model_epsilon_a.epsilon_r
+        a = self.mod_mu_b - self.mod_mu_a
+        b = self.mod_epsilon_b - self.mod_epsilon_a
+        assert_allclose(a.mu_r, self.mod_mu_a.mu_r)
+        assert_allclose(b.epsilon_r, self.mod_epsilon_a.epsilon_r)
 
     def test_eq(self):
 
-        assert (self.model_int == self.mesh_base) is False
+        assert (self.mod_iso == self.mesh_base) is False
 
-        out = self.model_int == self.model_int_diff
+        out = self.mod_iso == self.mod_int_diff
         assert not out
 
-        out = self.model_int == self.model_int
+        out = self.mod_iso == self.mod_iso
         assert out
 
     def test_general(self):
 
         # Check shape and size
-        assert_allclose(self.model_int.shape, self.mesh_base.shape_cells)
-        assert self.model_int.size == self.mesh_base.n_cells
+        assert_allclose(self.mod_iso.shape, self.mesh_base.shape_cells)
+        assert self.mod_iso.size == self.mesh_base.n_cells
 
     def test_copy(self):
-        model_new1 = self.model_shape_cells.copy()
-        model_new2 = self.model_3_a.copy()
-        model_new3 = self.model_mu_a.copy()
-        model_new4 = self.model_epsilon_a.copy()
+        model_new1 = self.mod_shape.copy()
+        model_new2 = self.mod_tri_a.copy()
+        model_new3 = self.mod_mu_a.copy()
+        model_new4 = self.mod_epsilon_a.copy()
 
-        assert model_new1 == self.model_shape_cells
-        assert model_new2 == self.model_3_a
-        assert model_new3 == self.model_mu_a
-        assert model_new4 == self.model_epsilon_a
+        assert model_new1 == self.mod_shape
+        assert model_new2 == self.mod_tri_a
+        assert model_new3 == self.mod_mu_a
+        assert model_new4 == self.mod_epsilon_a
 
         assert (model_new1.property_x.base is not
-                self.model_shape_cells.property_x.base)
-        assert model_new2.property_y.base is not self.model_3_a.property_y.base
-        assert model_new2.property_z.base is not self.model_3_a.property_z.base
-        assert model_new3.mu_r.base is not self.model_mu_a.mu_r.base
+                self.mod_shape.property_x.base)
+        assert model_new2.property_y.base is not self.mod_tri_a.property_y.base
+        assert model_new2.property_z.base is not self.mod_tri_a.property_z.base
+        assert model_new3.mu_r.base is not self.mod_mu_a.mu_r.base
         assert (model_new4.epsilon_r.base is not
-                self.model_epsilon_a.epsilon_r.base)
+                self.mod_epsilon_a.epsilon_r.base)
 
     def test_dict(self):
         # dict is already tested via copy. Just the other cases here.
-        mdict = self.model_3_b.to_dict()
+        mdict = self.mod_tri_b.to_dict()
         keys = ['property_x', 'property_y', 'property_z', 'mu_r', 'epsilon_r',
                 'grid']
         for key in keys:
             assert key in mdict.keys()
         for key in keys[:3]:
-            val = getattr(self.model_3_b, key)
+            val = getattr(self.mod_tri_b, key)
             assert_allclose(mdict[key], val)
 
         del mdict['property_x']